feat: more helpful funcs in ai-shared
